requirements.txt
django==4.2
psycopg2-binary==2.9.6
djangorestframework==3.14.0
python-dotenv==1.0.0
validate_docbr 
django-filter
